refactor: share vocabulary across games
